requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
alphashape==1.3.1
anndata==0.12.2
annotated-types==0.7.0
anyio==4.10.0
ascii_colors==0.11.4
asttokens==3.0.0
asyncpg==0.30.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.5
Brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
click-log==0.4.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.3
configparser==7.2.0
Crawl4AI==0.7.4
cryptography==46.0.1
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
dotenv==0.9.9
ecdsa==0.19.1
executing==2.2.1
fake-http-header==0.3.5
fake-useragent==2.2.0
fastapi==0.116.2
fastuuid==0.12.0
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.9.0
future==1.0.0
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.0
humanize==4.13.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.3.2
json_repair==0.50.1
jsonlines==4.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
lark==1.2.2
lightrag-hku==1.4.8.2
litellm==1.77.1
lxml==5.4.0
markdown-it-py==4.0.0
markdownify==1.2.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.6.4
nano-vectordb==0.0.4.3
neo4j==5.28.2
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numba==0.61.2
numpy==2.3.3
ollama==0.5.4
openai==1.108.0
packaging==25.0
pandas==2.3.2
parso==0.8.5
passlib==1.7.4
patchright==1.55.2
pillow==11.3.0
pipmaster==1.0.4
platformdirs==4.4.0
playwright==1.55.0
prompt_toolkit==3.0.52
propcache==0.3.2
psutil==7.1.0
pure_eval==0.2.3
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pynndescent==0.5.13
pyOpenSSL==25.3.0
pypinyin==0.55.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.1.0
rank-bm25==0.2.2
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
rtree==1.4.1
scanpy==1.11.4
scikit-misc==0.5.1
scipy==1.16.2
seaborn==0.13.2
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.8
spyder-kernels==2.4.4
stack-data==0.6.3
starlette==0.48.0
statsmodels==0.14.5
tenacity==9.1.2
tf-playwright-stealth==1.2.0
tiktoken==0.11.0
tokenizers==0.22.0
torchvision==0.23.0+cu126
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
trimesh==4.8.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
umap-learn==0.5.9.post2
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
xlsxwriter==3.2.9
xxhash==3.5.0
yarl==1.20.1
zarr==3.1.1
zipp==3.23.0
